display: Fix up kernel header dependencies declaration
Currently, when the kernel headers must be included
unconditionally, the dependency on them is not expressed
in the standard format.

For Android makefiles that must include the kernel headers,
declare the local additional dependencies unconditionally
and in standard form as well. Also, remove kernel_deps from
makefiles, as it is not used.

CRs-Fixed: 2439890
Change-Id: I928fdc6f734431fa0d60c1a3624b533a9d6a85b7